requirements.txt
# Solstis Voice Assistant Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
python-dotenv>=1.0.0

# OpenAI API for GPT-4o mini and TTS
openai>=1.0.0

# Picovoice for wake word detection
pvporcupine>=3.0.0

# Audio processing (optional - we use ALSA directly)
# pyaudio>=0.2.11

# System dependencies (install via package manager):
# - ALSA development libraries: sudo apt-get install libasound2-dev (Ubuntu/Debian)
# - Python development headers: sudo apt-get install python3-dev (Ubuntu/Debian)
